feat: test traffic whitelisting
License: MIT
