+++ plans/network/go.mod
@@ -2,4 +2,4 @@ module github.com/testground/testground/plans/network
 
 go 1.14
 
-require github.com/testground/sdk-go v0.2.2
+require github.com/testground/sdk-go v0.2.3-0.20200721144813-6d25ed275153
