+++ plans/network/manifest.toml
@@ -26,3 +26,13 @@ enabled = true
 [[testcases]]
 name = "ping-pong"
 instances = { min = 2, max = 2, default = 2 }
+
+# seq 1
+[[testcases]]
+name = "traffic-allowed"
+instances = { min = 1, max = 1, default = 1 }
+
+# seq 2
+[[testcases]]
+name = "traffic-blocked"
+instances = { min = 1, max = 1, default = 1 }
